+++ .release-please-manifest.json
@@ -8,5 +8,5 @@
   "src/OpenFeature.Contrib.Providers.Statsig": "0.1.0",
   "src/OpenFeature.Contrib.Providers.Flipt": "0.0.5",
   "src/OpenFeature.Contrib.Providers.EnvVar": "0.0.7",
-  "src/OpenFeature.Contrib.Providers.Ofrep": "0.0.1"
+  "src/OpenFeature.Providers.Ofrep": "0.0.1"
 }
