--- src/OpenFeature.Contrib.Providers.Ofrep/OpenFeature.Contrib.Providers.Ofrep.csproj
@@ -1,48 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <PackageId>OpenFeature.Contrib.Providers.Ofrep</PackageId>
-    <VersionNumber>0.0.1</VersionNumber> <!--x-release-please-version -->
-    <VersionPrefix>$(VersionNumber)</VersionPrefix>
-    <AssemblyVersion>$(VersionNumber)</AssemblyVersion>
-    <FileVersion>$(VersionNumber)</FileVersion>
-    <Description>OFREP provider for .NET</Description>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <EnableNETAnalyzers>true</EnableNETAnalyzers>
-    <NoWarn>EnableGenerateDocumentationFile</NoWarn>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="8.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="8.0.1" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
-    <PackageReference Include="Polly" Version="7.2.3" />
-    <PackageReference Include="System.Text.Json" Version="8.0.5" />
-    <PackageReference Include="OpenFeature" Version="$(OpenFeatureVersion)" />
-    <PackageReference Include="OpenFeature.DependencyInjection" Version="$(OpenFeatureVersion)" />
-  </ItemGroup>
-  <PropertyGroup>
-    <OpenFeatureVersion>[2.2,2.99999]</OpenFeatureVersion>
-  </PropertyGroup>
-
-  <ItemGroup
-    Condition=" '$(TargetFramework)' == 'net462' or '$(TargetFramework)' == 'netstandard2.0'">
-    <PackageReference Include="System.Net.Http" Version="4.3.4" />
-    <PackageReference Include="System.Net.Http.Json" Version="8.0.1" />
-    <PackageVersion Include="System.ValueTuple" Version="4.5.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <!-- make the internal methods visble to our test project -->
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-      <_Parameter1>$(MSBuildProjectName).Test</_Parameter1>
-    </AssemblyAttribute>
-  </ItemGroup>
-
-</Project>
