+++ src/OpenFeature.Providers.Ofrep/version.txt
