--- test/OpenFeature.Contrib.Providers.Ofrep.Test/OpenFeature.Contrib.Providers.Ofrep.Test.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <EnableNETAnalyzers>true</EnableNETAnalyzers>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference
-      Include="..\..\src\OpenFeature.Contrib.Providers.Ofrep\OpenFeature.Contrib.Providers.Ofrep.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Moq" Version="4.18.4" />
-  </ItemGroup>
-
-</Project>
